- changed double for salaries and int for district Ids to Long. It just works better with the database and better format when displaying info since we will likely see salaries with 5-6 figures. - added javadoc tags in case we create documentation

+++ calofficials/src/main/java/api/model/AssemblyMember.java
@@ -2,7 +2,8 @@ package api.model;
 
 /**
  * AssemblyMember class that extends Official.
- * April 19, 2026
+ * @author Kelvin Myat
+ * @since April 19, 2026
  */
 public class AssemblyMember extends Official {
 
@@ -16,7 +17,7 @@ public class AssemblyMember extends Official {
      * @param districtId the district id
      */
     public AssemblyMember(String name, String party, String email, String office,
-                          double salary, int districtId) {
+                          Long salary, Long districtId) {
         super(name, party, email, office, salary, districtId, "Assembly Member");
     }
 }
